requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
jinja2
python-multipart
groq
python-dotenv